+++ tests/src/freelist_test.c
@@ -9,7 +9,7 @@ static void test_freelist_create(void) {
   printf("  Running test_freelist_create...\n");
 
   VkrFreeList freelist;
-  const uint32_t TOTAL_SIZE = 1024;
+  const uint64_t TOTAL_SIZE = 1024;
 
   // Allocate memory for freelist nodes
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL_SIZE);
@@ -35,19 +35,19 @@ static void test_allocate_exact_and_refill(void) {
   printf("  Running test_allocate_exact_and_refill...\n");
 
   VkrFreeList freelist;
-  const uint32_t TOTAL_SIZE = 1024;
+  const uint64_t TOTAL_SIZE = 1024;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL_SIZE);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL_SIZE, &freelist));
 
-  uint32_t offset = VKR_INVALID_ID;
+  uint64_t offset = VKR_INVALID_ID;
   assert(vkr_freelist_allocate(&freelist, TOTAL_SIZE, &offset));
   assert(offset == 0);
   assert(vkr_freelist_free_space(&freelist) == 0);
   // Next allocation should fail
-  uint32_t off2 = VKR_INVALID_ID;
+  uint64_t off2 = VKR_INVALID_ID;
   assert(!vkr_freelist_allocate(&freelist, 1, &off2));
 
   // Free and allocate again
@@ -65,15 +65,15 @@ static void test_allocate_split_then_coalesce(void) {
   printf("  Running test_allocate_split_then_coalesce...\n");
 
   VkrFreeList freelist;
-  const uint32_t TOTAL_SIZE = 1024;
+  const uint64_t TOTAL_SIZE = 1024;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL_SIZE);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL_SIZE, &freelist));
 
-  uint32_t o1 = VKR_INVALID_ID;
-  uint32_t o2 = VKR_INVALID_ID;
+  uint64_t o1 = VKR_INVALID_ID;
+  uint64_t o2 = VKR_INVALID_ID;
   assert(vkr_freelist_allocate(&freelist, 200, &o1));
   assert(o1 == 0);
   assert(vkr_freelist_free_space(&freelist) == TOTAL_SIZE - 200);
@@ -99,14 +99,14 @@ static void test_multiple_alloc_free_patterns(void) {
   printf("  Running test_multiple_alloc_free_patterns...\n");
 
   VkrFreeList fr;
-  const uint32_t TOTAL = 2048;
+  const uint64_t TOTAL = 2048;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL, &fr));
 
-  uint32_t a = VKR_INVALID_ID, b = VKR_INVALID_ID, c = VKR_INVALID_ID,
+  uint64_t a = VKR_INVALID_ID, b = VKR_INVALID_ID, c = VKR_INVALID_ID,
            d = VKR_INVALID_ID;
   assert(vkr_freelist_allocate(&fr, 256, &a)); // [0..256)
   assert(vkr_freelist_allocate(&fr, 512, &b)); // [256..768)
@@ -133,14 +133,14 @@ static void test_double_free_detection(void) {
   printf("  Running test_double_free_detection...\n");
 
   VkrFreeList fr;
-  const uint32_t TOTAL = 1024;
+  const uint64_t TOTAL = 1024;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL, &fr));
 
-  uint32_t off = VKR_INVALID_ID;
+  uint64_t off = VKR_INVALID_ID;
   assert(vkr_freelist_allocate(&fr, 128, &off)); // [0..128)
   assert(vkr_freelist_free(&fr, 128, off));
   // Double free of the same range should be rejected
@@ -155,14 +155,14 @@ static void test_clear_resets_to_single_block(void) {
   printf("  Running test_clear_resets_to_single_block...\n");
 
   VkrFreeList fr;
-  const uint32_t TOTAL = 4096;
+  const uint64_t TOTAL = 4096;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL, &fr));
 
-  uint32_t o1 = VKR_INVALID_ID, o2 = VKR_INVALID_ID;
+  uint64_t o1 = VKR_INVALID_ID, o2 = VKR_INVALID_ID;
   assert(vkr_freelist_allocate(&fr, 512, &o1));
   assert(vkr_freelist_allocate(&fr, 512, &o2));
   assert(vkr_freelist_free_space(&fr) == TOTAL - 1024);
@@ -182,14 +182,14 @@ static void test_insert_into_empty_list(void) {
   printf("  Running test_insert_into_empty_list...\n");
 
   VkrFreeList fr;
-  const uint32_t TOTAL = 256;
+  const uint64_t TOTAL = 256;
 
   uint64_t mem_size = vkr_freelist_calculate_memory_requirement(TOTAL);
   void *memory = malloc(mem_size);
   assert(memory != NULL);
   assert(vkr_freelist_create(memory, mem_size, TOTAL, &fr));
 
-  uint32_t off = VKR_INVALID_ID;
+  uint64_t off = VKR_INVALID_ID;
   // Consume entire range
   assert(vkr_freelist_allocate(&fr, TOTAL, &off));
   assert(vkr_freelist_free_space(&fr) == 0);
